+++ cmdcompass/gui/commandbox.py
@@ -26,12 +26,30 @@ class CommandBox(ctk.CTkFrame):
         self.command_index = command_index
         self.main_window = main_window
 
+        # Frame for delete button
+        self.delete_button_frame = ctk.CTkFrame(self)
+        self.delete_button_frame.place(relx=0, rely=0, anchor="nw")  # Position relative to CommandBox
+
+        # Delete button
+        self.delete_button = ctk.CTkButton(
+            self.delete_button_frame,
+            text="x",  # Use a bullet point character or a small image
+            width=7,
+            height=7,
+            fg_color="red",
+            hover_color="#c77c78",
+            font=("TkDefaultFont ", 9),
+            command=lambda: self.main_window.delete_command(self.command_index)
+        )
+        self.delete_button.pack()
+        self.delete_button.configure(corner_radius=5)
+
         # Command label (first row)
         command_summery = command.command_str
-        if len(command_summery) > 25:
-            command_summery = command_summery[:25] + "..."
+        if len(command_summery) > 27:
+            command_summery = command_summery[:27] + "..."
         self.command_label = ctk.CTkLabel(self, text=command_summery,  font=("TkDefaultFont", 12))
-        self.command_label.grid(row=0, column=0, padx=10, pady=0, sticky="w")
+        self.command_label.grid(row=0, column=0, padx=(20,0), pady=0, sticky="w")
 
         # Tag box (second row, first column)
         tag_box = TagBox(self, tags, command, main_window)
@@ -45,7 +63,7 @@ class CommandBox(ctk.CTkFrame):
             height=40,
             command=lambda: self.main_window.on_command_select(self.command_index)
         )
-        self.select_button.grid(row=0, column=1, rowspan=2, padx=(2, 3), pady=2, sticky="ns")
+        self.select_button.grid(row=0, column=1, rowspan=2, padx=(0, 3), pady=2, sticky="ns")
 
         # Adjust column weights
         self.grid_columnconfigure(0, weight=1)